logits = net_output[0].float()

+++ fairseq/models/fairseq_decoder.py
@@ -87,7 +87,7 @@ class FairseqDecoder(nn.Module):
             out = self.adaptive_softmax.get_log_prob(net_output[0], target=target)
             return out.exp_() if not log_probs else out
 
-        logits = net_output[0]
+        logits = net_output[0].float()
         if log_probs:
             return utils.log_softmax(logits, dim=-1, onnx_trace=self.onnx_trace)
         else:
